pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "firebase-admin>=6.7.0",
    "httpx>=0.28.1",
    "orjson>=3.10.16",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.3",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
